requirements.txt
colorama==0.4.6
et-xmlfile==1.1.0
greenlet==3.1.1
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
XlsxWriter==3.2.0
